refactor/main: complete overhaul of is_gameover fn
I've created a new way of evaluation of winning state: check the
environment of each cell in the grid looking for a winning state. This
new check also evaluates diagonal lines and is a prototype that could be
reworked for a 4-in-line version of the game or even a n-in-line
version, with configurable board size.